add the acquisition time attribute special handling

+++ addons/dicom/dcm3d.cc
@@ -49,7 +49,7 @@ CDicom3DImageIOPlugin::CDicom3DImageIOPlugin():
 	add_supported_type(it_ushort);
 
 	TTranslator<float>::register_for("SliceLocation");
-	TTranslator<float>::register_for("AcquisitionTime"); 
+	TTranslator<double>::register_for("AcquisitionTime"); 
 	TTranslator<int>::register_for("SeriesNumber");
 	TTranslator<int>::register_for("AcquisitionNumber");
 	TTranslator<int>::register_for("InstanceNumber");
